Move WinAdapter into interface directory (#5066)
Since it is shipped and relied upon, the WinAdapter header belongs in
the release directory. This removes the need to include the Support
directory for building purposes

revises fix to #4918

(cherry picked from commit 5eb7c77d7c8ee0f513e27170c6fed855ac21de80)
