+++ include/dxc/DxilContainer/DxilRuntimeReflection.h
@@ -12,7 +12,7 @@
 #pragma once
 
 #include "dxc/DXIL/DxilConstants.h"
-#include "dxc/Support/WinAdapter.h"
+#include "dxc/WinAdapter.h"
 
 #define RDAT_NULL_REF ((uint32_t)0xFFFFFFFF)
