+++ include/dxc/HLSL/DxilValidation.h
@@ -13,7 +13,7 @@
 
 #include "dxc/Support/Global.h"
 #include "dxc/DXIL/DxilConstants.h"
-#include "dxc/Support/WinAdapter.h"
+#include "dxc/WinAdapter.h"
 #include <memory>
 
 namespace llvm {
