+++ include/dxc/Support/WinIncludes.h
@@ -59,7 +59,7 @@ template <class T> void swap(CComHeapPtr<T> &a, CComHeapPtr<T> &b) {
 
 #else // _MSC_VER
 
-#include "dxc/Support/WinAdapter.h"
+#include "dxc/WinAdapter.h"
 
 #ifdef __cplusplus
 // Define operator overloads to enable bit operations on enum values that are
