+++ include/dxc/Support/exception.h
@@ -11,7 +11,7 @@
 #pragma once
 
 #include "dxc/Support/ErrorCodes.h"
-#include "dxc/Support/WinAdapter.h"
+#include "dxc/WinAdapter.h"
 #include <exception>
 #include <string>
