+++ include/dxc/WinAdapter.h
