+++ lib/Support/regerror.c
@@ -45,7 +45,7 @@
 
 #include "regutils.h"
 
-#include "dxc/Support/WinAdapter.h" // HLSL Change
+#include "dxc/WinAdapter.h" // HLSL Change
 
 #ifdef _MSC_VER
 #define snprintf _snprintf
